% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_rows.R
\name{add_rows}
\alias{add_rows}
\title{Add rows to the table with summary statistics.}
\usage{
add_rows(table.data, row.vars, summary.stat = function(x) mean(x),
  digits = 2)
}
\arguments{
\item{table.data}{Output from the \code{outline_table} object.}

\item{row.vars}{The variables that you want added to the table. Must be from
the list supplied to \code{outline_table}.}

\item{summary.stat}{The summary statistic or any other function. A list of
built functions can be found in \code{\link{table_stats}}.}

\item{digits}{}
}
\value{
Outputs a \code{table_draft} object with the variables added to the
  rows.
}
\description{
Add rows to the table with summary statistics.
}
\seealso{
\code{\link{carpenter}} for a list of all functions, examples, and
  accessing the introduction tutorial vignette.
}

