% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carpenter.R
\name{carpenter}
\alias{carpenter}
\title{Build common tables for your research needs!}
\description{
Build common tables for your research needs!
}
\seealso{
\code{\link{add_rows}}, \code{\link{table_renaming}} for renaming of
  columns and rows, \code{\link{construct_table}},
  \code{\link{use_pixiedust}} and \code{\link{use_pander}} to for further
  refining with more detailed table packages, \code{\link{table_stats}} for a
  list of built-in summary statistics.
}

