% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_table.R
\name{construct_table}
\alias{construct_table}
\alias{construct_table.table_draft}
\title{Construct the final table.}
\usage{
construct_table(table.data, ...)

\method{construct_table}{table_draft}(table.data, caption = NULL,
  style = "rmarkdown", alignment = function(x) ifelse(sapply(x, is.numeric),
  "center", "left"), split = Inf, missing = "")
}
\arguments{
\item{table.data}{The \code{table_draft} object.}

\item{...}{Additional parameters for \code{link{pander}} options}
}
\value{
Creates a \code{\link{pander}} created table.
}
\description{
Output can be to common formats such as rmarkdown, html, etc, based on the
\code{style} argument of the \code{\link[pander]{pander}} function.
}
\seealso{
\code{\link{carpenter}} for a list of all functions, examples, and
  accessing the introduction tutorial vignette.
}

