% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline_table.R
\name{outline_table}
\alias{outline_table}
\title{Make an outline of the table you want to create.}
\usage{
outline_table(data, rows, columns)
}
\arguments{
\item{data}{Dataset to use to create the table}

\item{table.rows}{Column or variable(s) that will make up the rows}

\item{table.columns}{Column or variable(s) that will make up the columns}
}
\value{
Dataframe ready for further carpentry work, like adding rows,
  rounding, summary statistics, renaming, etc.
}
\description{
Make an outline of the table you want to create.
}
\seealso{
\code{\link{carpenter}} for a list of all functions, examples, and
  accessing the introduction tutorial vignette.
}

