% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renaming.R
\name{table_renaming}
\alias{rename_columns}
\alias{rename_columns.table_draft}
\alias{rename_columns.table_outline}
\alias{rename_rows}
\alias{rename_rows.table_draft}
\alias{rename_rows.table_outline}
\alias{table_renaming}
\title{Renaming row and column variables.}
\usage{
rename_rows(table.data, pattern = NULL)

rename_columns(table.data, ...)

\method{rename_rows}{table_outline}(table.data, pattern = NULL)

\method{rename_rows}{table_draft}(table.data, pattern.fun = NULL)

\method{rename_columns}{table_outline}(table.data, ...)

\method{rename_columns}{table_draft}(table.data, ...)
}
\arguments{
\item{table.data}{The \code{table_draft} object.}

\item{pattern}{Generally a function (anonymous or otherwise) using the
\code{\link[base]{gsub}} function to substitute patterns, words,
characters, or symbols, etc.}

\item{...}{Renamed column headers. See
\code{\link[pixiedust]{sprinkle_colnames}} for more help using this
argument.}
}
\value{
Outputs a \code{table_draft} object with columns or rows renamed,
  which will also keep the renaming even if more rows are added via
  \code{\link{add_rows}}.
}
\description{
Renaming row and column variables.
}
\seealso{
\code{\link{carpenter}} for a list of all functions, examples, and
  accessing the introduction tutorial vignette.
}

