% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{table_stats}
\alias{table_stats}
\alias{stat_median}
\alias{stat_iqr}
\alias{stat_medianIQR}
\alias{stat_mean}
\alias{stat_stddev}
\alias{stat_meanSD}
\alias{stat_nPct}
\title{Common summary statistics to use in \code{\link{add_rows}}.}
\usage{
stat_median(x, digits = 1)

stat_iqr(x, digits = 1)

stat_medianIQR(x, digits = 1)

stat_mean(x, digits = 1)

stat_stddev(x, digits = 1)

stat_meanSD(x, digits = 1)

stat_nPct(x, digits = 0)
}
\arguments{
\item{x}{Numeric vector to use to calculate the statistic}

\item{digits}{Number of digits to use}
}
\value{
Create a single character string with the summary statistic
}
\description{
Common summary statistics to use in \code{\link{add_rows}}.
}
\seealso{
\code{\link{carpenter}} for a list of all functions, examples, and
  accessing the introduction tutorial vignette.
}
