% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{use_pander}
\alias{use_pander}
\title{Ready for refining using pander.}
\usage{
use_pander(table.data)
}
\arguments{
\item{table.data}{The \code{table_draft} object.}
}
\value{
A table ready to be used by \code{\link[pander]{pander}}.
}
\description{
A thin wrapper to prepare the output for use in the \code{\link[pander]{pander}} function.
}
\seealso{
\code{\link{carpenter}} for a list of all functions, examples, and
  accessing the introduction tutorial vignette.
}

