% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{use_pixiedust}
\alias{use_pixiedust}
\title{Ready for pixiedusting.}
\usage{
use_pixiedust(table.data)
}
\arguments{
\item{table.data}{The \code{table_draft} object.}
}
\value{
A \code{\link[pixiedust]{dust}}'ed table ready to be used by
  \code{\link{pixiedust}}.
}
\description{
A thin wrapper around the \code{\link[pixiedust]{dust}} function.
}
\seealso{
\code{\link{carpenter}} for a list of all functions, examples, and
  accessing the introduction tutorial vignette.
}

