% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_rows.R
\name{add_rows}
\alias{add_rows}
\title{Add rows to the table with summary statistics.}
\usage{
add_rows(data, row_vars, stat, digits = 1)
}
\arguments{
\item{data}{Output from the \code{outline_table} object.}

\item{row_vars}{The variables that you want added to the table. Must be from
\code{outline_table}.}

\item{stat}{The summary statistic or any other function. A list of
built functions can be found in \code{\link{table_stats}}.}

\item{digits}{What to round the value to.}
}
\value{
Adds a row with summary statistics for a variable.
}
\description{
Add rows to the table with summary statistics.
}
\seealso{
\code{\link{carpenter}} for a list of all functions, examples, and
  accessing the introduction tutorial vignette. See \code{\link{table_stats}}
  for a list of carpenter builtin statistics.
}

